+++ pyzoo/zoo/examples/orca/automl/AutoXGBoostClassifier.py
@@ -21,7 +21,7 @@ import pandas as pd
 from sklearn.model_selection import train_test_split
 from zoo import init_spark_on_local, init_spark_on_yarn
 from zoo.ray import RayContext
-from zoo.orca.automl.AutoXGBoost import AutoXGBoost
+from zoo.orca.automl.xgboost import AutoXGBoost
 from zoo.zouwu.config.recipe import *
 
 if __name__ == '__main__':
