+++ pyzoo/zoo/examples/orca/automl/AutoXGBoostRegressor.py
@@ -19,7 +19,7 @@ import argparse
 from sklearn.model_selection import train_test_split
 from zoo import init_spark_on_local, init_spark_on_yarn
 from zoo.ray import RayContext
-from zoo.orca.automl.AutoXGBoost import AutoXGBoost
+from zoo.orca.automl.xgboost import AutoXGBoost
 from zoo.zouwu.config.recipe import *
 
 if __name__ == '__main__':
